Move CollectionTypes into Types
