+++ src/base/CollectionStateStore.hh
@@ -8,7 +8,6 @@
 #pragma once
 
 #include "Assert.hh"
-#include "CollectionTypes.hh"
 #include "OpaqueId.hh"
 #include "Types.hh"
