--- src/base/CollectionTypes.hh
@@ -1,34 +1,0 @@
-//----------------------------------*-C++-*----------------------------------//
-// Copyright 2021 UT-Battelle, LLC, and other Celeritas developers.
-// See the top-level COPYRIGHT file for details.
-// SPDX-License-Identifier: (Apache-2.0 OR MIT)
-//---------------------------------------------------------------------------//
-//! \file CollectionTypes.hh
-//---------------------------------------------------------------------------//
-#pragma once
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-//! Memory location of data
-enum class MemSpace
-{
-    host,
-    device,
-#ifdef __CUDACC__
-    native = device, // Included by a CUDA file
-#else
-    native = host,
-#endif
-};
-
-//! Data ownership flag
-enum class Ownership
-{
-    value,           //!< Ownership of the data, only on host
-    reference,       //!< Mutable reference to the data
-    const_reference, //!< Immutable reference to the data
-};
-
-//---------------------------------------------------------------------------//
-} // namespace celeritas
